server/requirements.txt
Django
psycopg2-binary
djangorestframework
drf-yasg
django-debug-toolbar
uvicorn
gunicorn